# Conflicts:
#	src/main/java/ch/uzh/ifi/hase/soprafs23/controller/UserController.java

+++ src/main/java/ch/uzh/ifi/hase/soprafs23/controller/UserController.java
@@ -1,5 +1,7 @@
 package ch.uzh.ifi.hase.soprafs23.controller;
 
+import ch.uzh.ifi.hase.soprafs23.constant.GroupState;
+import ch.uzh.ifi.hase.soprafs23.entity.Group;
 import ch.uzh.ifi.hase.soprafs23.entity.Invitation;
 import ch.uzh.ifi.hase.soprafs23.entity.User;
 import ch.uzh.ifi.hase.soprafs23.rest.dto.GroupGetDTO;
@@ -132,12 +134,38 @@ public class UserController {
             throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You are not authorized.");
         }
 
+        // Check if the user is in a group that is not in the GROUPFORMING state
+        Group currentGroup = groupService.getGroupByUserId(userId);
+        if(currentGroup != null && currentGroup.getGroupState() != GroupState.GROUPFORMING) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "You cannot logout while your group is beyond the forming stage.");
+        }
+
         // delete all join requests of the user
         joinRequestService.deleteAllJoinRequests(userId);
 
+        // Check if the user is a host of a group
+        Group hostedGroup = null;
+        try {
+            hostedGroup = groupService.getGroupByHostId(userId);
+            if (hostedGroup != null) {
+                // delete all the invitations that were sent out for this group
+                invitationService.deleteInvitationsByGroupId(hostedGroup.getId());
+
+                // delete all the join requests to join this group
+                joinRequestService.deleteJoinRequestsByGroupId(hostedGroup.getId());
+
+                // delete the group
+                groupService.deleteGroup(hostedGroup.getId());
+            }
+        } catch (ResponseStatusException ex) {
+            // If the user is not a host, they might still be a member of a group
+            if(userService.isUserInGroup(userId)) {
+                userService.leaveGroup(userId);
+            }
+        }
+
         userService.logout(userId);
     }
-
 
     @PutMapping("/users/{userId}")
     @ResponseStatus(HttpStatus.NO_CONTENT) //NO CONTENT IS 204
